image-to-pdf-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
img2pdf==0.5.1
python-dotenv==1.0.0
aiofiles==23.2.1
pypdf==3.17.4
pdf2docx==0.5.8
pdfplumber==0.11.4
openpyxl==3.1.2
python-pptx==0.6.23
pdf2image==1.17.0
pikepdf==8.11.2
